Backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
distlib==0.3.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==3.5.0
Pillow==8.3.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.2
zipp==3.4.0
